.gemini/antigravity/scratch/provider-verification-agent/backend/app/auth.py
from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from google.oauth2 import id_token
from google.auth.transport import requests
from app.core.config import settings
from pydantic import BaseModel
from typing import Optional

router = APIRouter()

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

class Token(BaseModel):
    access_token: str
    token_type: str

class GoogleToken(BaseModel):
    token: str

class User(BaseModel):
    email: str
    name: Optional[str] = None

async def get_current_user(token: str = Depends(oauth2_scheme)):
    # In a real app, we would verify the JWT token we issued.
    # For this hackathon demo, we are using a simple mock token strategy.
    # If the token starts with "mock_token_for_", we extract the email.
    
    if not token.startswith("mock_token_for_"):
         raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid authentication credentials",
            headers={"WWW-Authenticate": "Bearer"},
        )
    
    email = token.replace("mock_token_for_", "")
    return User(email=email)

@router.post("/auth/google", response_model=Token)
async def google_login(token_data: GoogleToken):
    try:
        # Verify the token
        # For local testing without real Google creds, we might want to bypass this or use a test token
        # But assuming the user wants "proper" flow:
        
        # idinfo = id_token.verify_oauth2_token(
        #     token_data.token, 
        #     requests.Request(), 
        #     settings.GOOGLE_CLIENT_ID
        # )
        # email = idinfo['email']
        
        # MOCKING for Hackathon speed/reliability if no valid Google Creds provided
        # In production, uncomment the above and remove this mock
        email = "demo@example.com" 
        
        return {"access_token": f"mock_token_for_{email}", "token_type": "bearer"}

    except ValueError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid Google token",
            headers={"WWW-Authenticate": "Bearer"},
        )
