.gemini/antigravity/scratch/provider-verification-agent/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
beautifulsoup4
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
